Working on manga

+++ Interfaces/IData.cs
@@ -10,11 +10,11 @@ namespace WeebLib.Interfaces
     {
         public string name { get; private set; }
         public int totalChapters { get; private set; }
-        public int latestChapter { get; private set; }
+        public double latestChapter { get; private set; }
         internal string initalLink { get; private set; }
         public string source { get; private set; }
 
-        public IData(string name, int totalChapterNum, int latestChapter, string initalLink, string source)
+        public IData(string name, int totalChapterNum, double latestChapter, string initalLink, string source)
         {
             this.name = name;
             this.totalChapters = totalChapterNum;
